src/backend/pyproject.toml
[project]
name = "subbyx-backend"
version = "0.1.0"
description = "Subbyx Fraud Detection API"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pandas>=2.2.0",
    "pyarrow>=14.0.0",
    "pydantic>=2.5.3",
    "python-multipart>=0.0.6",
    "mlflow>=2.8.0",
    "lightgbm>=4.0.0",
    "scikit-learn>=1.4.0",
    "joblib>=1.3.0",
    "feast[redis]>=0.40.0",
    "pyyaml>=6.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.1.0",
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]

[tool.ruff]
line-length = 100
exclude = [
    ".venv",
    "__pycache__",
    ".ruff_cache",
    "*.pyc",
    ".git",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff.format]
quote-style = "double"
